Revert "readme edit"
This reverts commit d22f03488c7b82ab115360098c9510e2a74b5c8b.

+++ projects/bridge-lz/contracts/abstract/AbstractCrossChainAdapter.sol
@@ -46,7 +46,7 @@ abstract contract AbstractCrossChainAdapter is ICrossChainBridge {
     }
 
     //primary function for receiving ETH from other chain
-    function _handleCrossChainEth(uint32 _chainId) internal {
+    function _handleCrossChainEth(uint256 _chainId) internal {
         emit CrossChainEthDeposit(_chainId, msg.value);
         Address.sendValue(payable(targetReceiver), msg.value);
     }
